Consolidate FormBuilder into a single file
Currently , the FormBuilder class is defined in the FormHelper class and then reopened and complemented in the FormOptionsHelper class.

This PR has been created because understanding all the moving pieces for FormBuilders, FormHelpers and FormOptionHelpers becomes increasingly hard when classes are created and reopened in multiple places. Consolidating it into a single file makes it easier to understand the breadth and purpose of the class.

This PR is part of a larger effort to cleanup and modernise this section of the codebase to hopefully bring in newer abstractions to it. See #48574 for a previous PR with the same purpose.
