+++ actionview/lib/action_view/helpers/form_options_helper.rb
@@ -835,91 +835,5 @@ module ActionView
           prompt.kind_of?(String) ? prompt : I18n.translate("helpers.select.prompt", default: "Please select")
         end
     end
-
-    class FormBuilder
-      # Wraps ActionView::Helpers::FormOptionsHelper#select for form builders:
-      #
-      #   <%= form_for @post do |f| %>
-      #     <%= f.select :person_id, Person.all.collect { |p| [ p.name, p.id ] }, include_blank: true %>
-      #     <%= f.submit %>
-      #   <% end %>
-      #
-      # Please refer to the documentation of the base helper for details.
-      def select(method, choices = nil, options = {}, html_options = {}, &block)
-        @template.select(@object_name, method, choices, objectify_options(options), @default_html_options.merge(html_options), &block)
-      end
-
-      # Wraps ActionView::Helpers::FormOptionsHelper#collection_select for form builders:
-      #
-      #   <%= form_for @post do |f| %>
-      #     <%= f.collection_select :person_id, Author.all, :id, :name_with_initial, prompt: true %>
-      #     <%= f.submit %>
-      #   <% end %>
-      #
-      # Please refer to the documentation of the base helper for details.
-      def collection_select(method, collection, value_method, text_method, options = {}, html_options = {})
-        @template.collection_select(@object_name, method, collection, value_method, text_method, objectify_options(options), @default_html_options.merge(html_options))
-      end
-
-      # Wraps ActionView::Helpers::FormOptionsHelper#grouped_collection_select for form builders:
-      #
-      #   <%= form_for @city do |f| %>
-      #     <%= f.grouped_collection_select :country_id, @continents, :countries, :name, :id, :name %>
-      #     <%= f.submit %>
-      #   <% end %>
-      #
-      # Please refer to the documentation of the base helper for details.
-      def grouped_collection_select(method, collection, group_method, group_label_method, option_key_method, option_value_method, options = {}, html_options = {})
-        @template.grouped_collection_select(@object_name, method, collection, group_method, group_label_method, option_key_method, option_value_method, objectify_options(options), @default_html_options.merge(html_options))
-      end
-
-      # Wraps ActionView::Helpers::FormOptionsHelper#time_zone_select for form builders:
-      #
-      #   <%= form_for @user do |f| %>
-      #     <%= f.time_zone_select :time_zone, nil, include_blank: true %>
-      #     <%= f.submit %>
-      #   <% end %>
-      #
-      # Please refer to the documentation of the base helper for details.
-      def time_zone_select(method, priority_zones = nil, options = {}, html_options = {})
-        @template.time_zone_select(@object_name, method, priority_zones, objectify_options(options), @default_html_options.merge(html_options))
-      end
-
-      # Wraps ActionView::Helpers::FormOptionsHelper#weekday_select for form builders:
-      #
-      #   <%= form_for @user do |f| %>
-      #     <%= f.weekday_select :weekday, include_blank: true %>
-      #     <%= f.submit %>
-      #   <% end %>
-      #
-      # Please refer to the documentation of the base helper for details.
-      def weekday_select(method, options = {}, html_options = {})
-        @template.weekday_select(@object_name, method, objectify_options(options), @default_html_options.merge(html_options))
-      end
-
-      # Wraps ActionView::Helpers::FormOptionsHelper#collection_check_boxes for form builders:
-      #
-      #   <%= form_for @post do |f| %>
-      #     <%= f.collection_check_boxes :author_ids, Author.all, :id, :name_with_initial %>
-      #     <%= f.submit %>
-      #   <% end %>
-      #
-      # Please refer to the documentation of the base helper for details.
-      def collection_check_boxes(method, collection, value_method, text_method, options = {}, html_options = {}, &block)
-        @template.collection_check_boxes(@object_name, method, collection, value_method, text_method, objectify_options(options), @default_html_options.merge(html_options), &block)
-      end
-
-      # Wraps ActionView::Helpers::FormOptionsHelper#collection_radio_buttons for form builders:
-      #
-      #   <%= form_for @post do |f| %>
-      #     <%= f.collection_radio_buttons :author_id, Author.all, :id, :name_with_initial %>
-      #     <%= f.submit %>
-      #   <% end %>
-      #
-      # Please refer to the documentation of the base helper for details.
-      def collection_radio_buttons(method, collection, value_method, text_method, options = {}, html_options = {}, &block)
-        @template.collection_radio_buttons(@object_name, method, collection, value_method, text_method, objectify_options(options), @default_html_options.merge(html_options), &block)
-      end
-    end
   end
 end
